+++ modules/drink_store/drink_store-api/src/main/java/com/ntthanh/drink_store/service/DrinkServiceWrapper.java
@@ -77,7 +77,7 @@ public class DrinkServiceWrapper
 
 	@Override
 	public void updateDrink(
-		long drinkId, String drinkName, String category, long price,
+		long drinkId, String drinkName, String category, double price,
 		String imageUrl) {
 
 		_drinkService.updateDrink(
