+++ modules/drink_store/drink_store-service/src/main/java/com/ntthanh/drink_store/service/impl/DrinkServiceImpl.java
@@ -49,7 +49,7 @@ public class DrinkServiceImpl extends DrinkServiceBaseImpl {
 	}
 
 	@Override
-	public void updateDrink(long drinkId, String drinkName, String category, long price, String imageUrl) {
+	public void updateDrink(long drinkId, String drinkName, String category, double price, String imageUrl) {
 		try {
 			Drink drink = drinkLocalService.getDrink(drinkId);
